Move all viewer files into the editor folder, and merge footer and header files
