+++ novelwriter/editor/viewerpanel.py
